chore: general improvements and refactors of headless-ssr-commerce (#4625)
https://coveord.atlassian.net/browse/KIT-3681

Some general improvements before I added the cart.

- moved the components, lib and utils folder back to root so we follow
nextjs conventions & the imports are cleaner like so :
```
import Triggers from '@/components/triggers/triggers';
import {searchEngineDefinition} from '@/lib/commerce-engine';
```
- renamed the SearchPage and ListingPage as SearchProvider and
ListingProvider as this is what they do currently.

- Fix the css layout of the pages so the facet don't take 95% of the
space

---------
