+++ packages/headless-react/package.json
@@ -39,7 +39,6 @@
   "devDependencies": {
     "@coveo/release": "1.0.0",
     "@testing-library/react": "14.3.1",
-    "@types/jest": "29.5.12",
     "@typescript-eslint/eslint-plugin": "7.17.0",
     "eslint-plugin-react": "7.35.0",
     "eslint-plugin-testing-library": "6.2.2",
