+++ packages/samples/headless-ssr-commerce/components/facets/category-facet.tsx
