+++ packages/samples/headless-ssr-commerce/components/facets/numeric-facet.tsx
