+++ packages/samples/headless-ssr-commerce/components/facets/regular-facet.tsx
