+++ packages/samples/headless-ssr-commerce/utils/format-currency.ts
